casi FINAL

+++ src/main/java/Modelo/Line.java
@@ -9,31 +9,31 @@ package Modelo;
  * @author santi
  */
 public class Line {
-    private int x1;
-    private int y1;
-    private int x2;
-    private int y2;
+    private double x1;
+    private double y1;
+    private double x2;
+    private double y2;
 
-    public Line(int x1, int y1, int x2, int y2) {
+    public Line(double x1, double y1, double x2, double y2) {
         this.x1 = x1;
         this.y1 = y1;
         this.x2 = x2;
         this.y2 = y2;
     }
 
-    public int getX1() {
+    public double getX1() {
         return x1;
     }
 
-    public int getY1() {
+    public double getY1() {
         return y1;
     }
 
-    public int getX2() {
+    public double getX2() {
         return x2;
     }
 
-    public int getY2() {
+    public double getY2() {
         return y2;
     }
